# doc/api.yaml
openapi: 3.1.0
info:
  title: VaikaParts Payment Backend
  description: >
    Payment backend for the VaikaParts application. This service is responsible for handling all payment operations within the VaikaParts platform, including merchant pay transactions via MVola, Orange Money, and Airtel Money. It operates independently from the VaikaParts domain backend, which handles the core business logic of the application.

  version: 1.0.0
  contact:
    name: Devik'Apps Development Team
    email: vaikaparts@gmail.com
    url: https://devikapps.com
servers:
  - url: https://xxxxxxx.com
    description: Pre-production environment
  - url: http://localhost:9090
    description: Local development environment
paths:
  /:
    get:
      operationId: addViewControllers
      tags:
        - Docs
      summary: Redirect root path to Swagger UI
      responses:
        302:
          description: Redirects to /doc
          headers:
            Location:
              description: URL to redirect
              schema:
                type: string
                format: uri
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServer'
  /ping:
    get:
      operationId: ping
      tags:
        - Health
      summary: Check if the server is alive
      responses:
        200:
          description: A message showing that the server is alive
          content:
            text/plain:
              schema:
                type: string
                format: string
                example: pong
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServer'
  /health/email:
    get:
      operationId: sendHealthCheckEmails
      tags:
        - Health
      summary: Send health check email
      description: Sends a test email to verify email service functionality
      security: []
      parameters:
        - name: to
          in: query
          required: true
          description: Email address to send the health check email to
          schema:
            type: string
            format: email
      responses:
        200:
          description: Email sent successfully
          content:
            text/plain:
              schema:
                type: string
                format: string
                example: "Email sent successfully"
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServer'
  /health/message:
    get:
      operationId: triggerDummyEvents
      tags:
        - Health
      summary: Trigger dummy health check events
      description: >
        This endpoint triggers one or more dummy events through the event producer  for testing the messaging system (e.g., RabbitMQ or Kafka).

      parameters:
        - name: nbEvent
          in: query
          required: false
          description: Number of events to trigger (1–500)
          schema:
            type: integer
            default: 1
            minimum: 1
            maximum: 500
        - name: waitInSeconds
          in: query
          required: false
          description: Duration (in seconds) to simulate event processing
          schema:
            type: integer
            default: 2
            minimum: 1
      responses:
        200:
          description: List of UUIDs corresponding to triggered dummy events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UUID'
        400:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/InternalServer'
  /health/db:
    get:
      operationId: checkDbHealth
      summary: Health check for the dummy database
      description: Returns a paginated list of Dummy entities.
      tags:
        - Health
      parameters:
        - name: page
          $ref: '#/components/parameters/Page'
        - name: size
          $ref: '#/components/parameters/PageSize'
      responses:
        200:
          description: Successful database health check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DummyPageResponse"
        400:
          $ref: '#/components/responses/BadRequest'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/payments/mvola:
    post:
      tags:
        - MVola
      summary: Initiate a payment
      description: Initiates a new MVola merchant pay transaction. The payment is created with a PENDING status. The caller must poll GET /{transactionId} or wait for the callback to know the final status.
      operationId: initiateMvolaPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MvolaPaymentRequest'
      responses:
        '201':
          description: Payment successfully initiated. Status is always PENDING at this stage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MvolaPayment'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/payments/mvola/customer/{customer-msisdn}:
    get:
      tags:
        - MVola
      summary: Get all Mvola payment by to a customer MSISDN.
      operationId: getMvolaPaymentsByCustomerMsisdn
      description: Retrieves all existing mvola payment transaction done by a given Customer MSISDN with possibility to retrieve with a given page and size.
      parameters:
        - $ref: '#/components/parameters/CustomerMsisdn'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        200:
          description: List of pagined Mvola payment done by the Customer with given MSISDN.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MvolaPaymentPageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/payments/mvola/{transactionId}:
    get:
      tags:
        - MVola
      summary: Get a payment by transaction ID
      description: Retrieves an existing MVola payment by its transaction ID. The transaction ID is the serverCorrelationId returned by the initiate endpoint.
      operationId: getMvolaPayment
      parameters:
        - name: transactionId
          in: path
          required: true
          description: The transaction ID of the payment to retrieve. This is the serverCorrelationId returned by the initiate endpoint.
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Payment found and returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MvolaPayment'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/payments/mvola/callback:
    put:
      tags:
        - MVola
      summary: Receive MVola payment callback
      description: >
        Endpoint invoked by MVola via PUT when a transaction reaches a terminal status (completed or failed). This endpoint is not called by the client — it is called by MVola's infrastructure only. On success, returns HTTP 200 with no body.

      operationId: handleMvolaCallBack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MvolaCallBackRequest'
      responses:
        '200':
          description: Callback received and processed successfully. No response body.
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/payments/airtel-money:
    post:
      tags:
        - Airtel Money
      summary: Initiate a payment
      description: Initiates a new Airtel Money merchant pay transaction. The payment is created with a PENDING status. The caller must poll GET /{transactionId} or wait for the callback to know the final status.
      operationId: initiateAirtelMoneyPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AirtelMoneyPaymentRequest'
      responses:
        '201':
          description: Payment successfully initiated. Status is always PENDING at this stage.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirtelMoneyPayment'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/payments/airtel-money/customer/{customer-msisdn}:
    get:
      tags:
        - Airtel Money
      summary: Get all Airtel Money payment by to a customer MSISDN.
      operationId: getAirtelMoneyPaymentsByCustomerMsisdn
      description: Retrieves all existing Airtel Money payment transaction done by a given Customer MSISDN with possibility to retrieve with a given page and size.
      parameters:
        - $ref: '#/components/parameters/CustomerMsisdn'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        200:
          description: List of pagined Airtel Money payment done by the Customer with given MSISDN.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirtelMoneyPaymentPageResponse'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/payments/airtel-money/{transactionId}:
    get:
      tags:
        - Airtel Money
      summary: Get a payment by transaction ID
      description: Retrieves an existing Airtel Money payment by its transaction ID.
      operationId: getAirtelMoneyPayment
      parameters:
        - name: transactionId
          in: path
          required: true
          description: The transaction ID of the payment to retrieve.
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Payment found and returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirtelMoneyPayment'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServer'
  /v1/payments/airtel-money/callback:
    put:
      tags:
        - Airtel Money
      summary: Receive Airtel Money payment callback
      description: >
        Endpoint invoked by Airtel Money via PUT when a transaction reaches a terminal status (completed or failed). This endpoint is not called by the client — it is called by Airtel Money's infrastructure only. On success, returns HTTP 200 with no body.

      operationId: handleAirtelMoneyCallBack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AirtelMoneyCallBackRequest'
      responses:
        '200':
          description: Callback received and processed successfully. No response body.
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServer'
components:
  parameters:
    CustomerMsisdn:
      name: customer-msisdn
      in: path
      required: true
      description: The customer msisdn with what the Mvola payment was done.
      schema:
        $ref: '#/components/schemas/PhoneNumber'
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
      description: Page number (0-based)
    PageSize:
      name: size
      in: query
      schema:
        type: integer
        default: 10
      description: Page size
  schemas:
    DateTime:
      type: string
      format: date-time
      description: Date and time in ISO 8601 format (RFC 3339)
      example: '2025-12-06T11:23:45Z'
    UUID:
      type: string
      pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      format: uuid
      example: e6b5cc3e-eaaa-4cb3-912c-e059e983bb78
    PhoneNumber:
      type: string
      pattern: '^\\+?[1-9]\\d{1,14}$'
      description: "E.164 phone number format"
      example: +261340000003
    Dummy:
      type: object
      required:
        - id
        - description
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        description:
          type: string
    PageResponse:
      type: object
      required: [content, pageable, total_elements, total_pages, last, first, size, number, number_of_elements, empty]
      description: >
        Paginated response wrapper matching the JSON serialization of org.springframework.data.domain.Page. The pagination metadata is nested under the `pageable` object, consistent with Spring Data defaults.

      properties:
        content:
          type: array
          description: Page content — the list of elements for the current page.
          items: {}
        pageable:
          type: object
          description: Pagination state used to produce this page.
          properties:
            page_number:
              type: integer
              description: Zero-based current page index.
            page_size:
              type: integer
              description: Requested number of items per page.
            offset:
              type: integer
              format: int64
              description: Offset of the first element of this page relative to the full result set.
            sort:
              $ref: '#/components/schemas/SortDescriptor'
            paged:
              type: boolean
              description: Whether pagination is applied.
            unpaged:
              type: boolean
              description: Whether pagination is disabled.
        total_elements:
          type: integer
          format: int64
          description: Total number of elements across all pages.
        total_pages:
          type: integer
          description: Total number of pages available.
        last:
          type: boolean
          description: Whether this is the last page.
        first:
          type: boolean
          description: Whether this is the first page.
        size:
          type: integer
          description: Effective page size (number of elements requested per page).
        number:
          type: integer
          description: Zero-based index of the current page.
        number_of_elements:
          type: integer
          description: Number of elements on the current page.
        empty:
          type: boolean
          description: Whether the current page has no content.
        sort:
          $ref: '#/components/schemas/SortDescriptor'
    SortDescriptor:
      type: object
      description: Sort state descriptor returned by Spring Data.
      properties:
        empty:
          type: boolean
          description: Whether no sort orders are applied.
        sorted:
          type: boolean
          description: Whether at least one sort order is applied.
        unsorted:
          type: boolean
          description: Whether no sort orders are applied (inverse of sorted).
    DummyPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Dummy'
    MvolaPaymentPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/MvolaPayment'
    AirtelMoneyPaymentPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/AirtelMoneyPayment'
    BasePayment:
      type: object
      description: Base payment domain model shared across all payment providers.
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        transaction_id:
          type: string
          example: "b7c3d4e5-f6a7-8901-bcde-f12345678901"
        description:
          type: string
          example: "Profile unlock payment"
        payer:
          $ref: '#/components/schemas/PaymentParty'
        payee:
          $ref: '#/components/schemas/PaymentParty'
        provider:
          $ref: '#/components/schemas/PaymentProvider'
        type:
          $ref: '#/components/schemas/PaymentType'
        amount:
          type: number
          format: decimal
          example: 5000
        currency:
          $ref: '#/components/schemas/PaymentCurrency'
        status:
          $ref: '#/components/schemas/VerificationStatus'
        created_at:
          $ref: '#/components/schemas/DateTime'
        updated_at:
          $ref: '#/components/schemas/DateTime'
    MvolaNotificationMethod:
      type: string
      enum: ['polling', 'callback']
      example: 'polling'
    MvolaPayment:
      type: obeject
      allOf:
        - $ref: '#/components/schemas/BasePayment'
      properties:
        server_correlation_id:
          $ref: '#/components/schemas/UUID'
        mvola_transaction_id:
          type: string
          example: "b7c3d4e5-f6a7-8901-bcde-f12345678901"
        notification_method:
          $ref: '#/components/schemas/MvolaNotificationMethod'
    AirtelMoneyPayment:
      type: object
      allOf:
        - $ref: '#/components/schemas/BasePayment'
      properties:
        airtel_money_id:
          $ref: '#/components/schemas/UUID'
    BasePaymentRequest:
      type: object
      description: Base payment request shared across all payment providers.
      required:
        - amount
        - payer
        - payee
        - type
      properties:
        amount:
          type: number
          format: decimal
          minimum: 0.01
          example: 5000
        currency:
          $ref: '#/components/schemas/PaymentCurrency'
        description:
          type: string
          example: "Profile unlock payment"
        payer:
          $ref: '#/components/schemas/PaymentParty'
        payee:
          $ref: '#/components/schemas/PaymentParty'
        type:
          $ref: '#/components/schemas/PaymentType'
    MvolaPaymentRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/BasePaymentRequest'
    AirtelMoneyPaymentRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/BasePaymentRequest'
    AirtelMoneyCAllBackTransaction:
      type: object
      description: This is the transaction payload sent by Airtel Money as part of the request body for callback url
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        message:
          description: This is the description of the transaction
          type: string
          example: Random description
        status_code:
          description: This is the Airtel transaction status code.
          type: string
          enum: [TS, TF]
          example: TS
        airtel_money_id:
          description: This is the airtel money system-generated transaction id
          type: string
          example: XDFI-D145DF
    AirtelMoneyCallBackRequest:
      type: object
      description: Callback payload sent by Airtel Money via POST when a transaction completes or fails
      properties:
        transaction:
          $ref: '#/components/schemas/AirtelMoneyCAllBackTransaction'
        hash:
          description: HMAC hash sent by Airtel when callback authentication is enabled. Null when callback authentication is disabled.
          type: string
          example: 968c2349040273dd57dc4be7e238c5ac200ceac5
    MvolaCallBackRequest:
      type: object
      description: Callback payload sent by MVola via PUT when a transaction completes or fails.
      properties:
        transactionStatus:
          type: string
          enum:
            - completed
            - failed
          example: "completed"
        serverCorrelationId:
          type: string
          format: uuid
          example: "421a22a2-ef1d-42bc-9452-f4939a3d5cdf"
        transactionReference:
          type: string
          example: "641235"
        requestDate:
          type: string
          format: date-time
          example: "2021-02-24T03:28:00.567Z"
        debitParty:
          type: array
          items:
            $ref: '#/components/schemas/MvolaPartyEntry'
        creditParty:
          type: array
          items:
            $ref: '#/components/schemas/MvolaPartyEntry'
        fees:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/MvolaFeeEntry'
        metadata:
          type: array
          items:
            $ref: '#/components/schemas/MvolaPartyEntry'
    MvolaPartyEntry:
      type: object
      description: Generic key-value entry used for debitParty, creditParty, and metadata in MVola payloads.
      properties:
        key:
          type: string
          example: "msisdn"
        value:
          type: string
          example: "0343500003"
    MvolaFeeEntry:
      type: object
      description: Fee entry returned in successful MVola callback payloads.
      properties:
        feeAmount:
          type: string
          example: "5.46"
    PaymentParty:
      type: object
      properties:
        phone_number:
          $ref: '#/components/schemas/PhoneNumber'
        name:
          type: string
          example: "Test Customer"
        country:
          type: string
          example: "MADAGASCAR"
    PaymentProvider:
      type: string
      enum:
        - MVOLA
        - ORANGE_MONEY
        - AIRTEL_MONEY
      example: "MVOLA"
    PaymentType:
      type: string
      enum:
        - PROFILE_UNLOCK
      example: "PROFILE_UNLOCK"
    PaymentCurrency:
      type: string
      enum:
        - AR
      example: "AR"
    VerificationStatus:
      type: string
      enum:
        - PENDING
        - SUCCESS
        - FAILED
      example: "PENDING"
    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      description: Standard error response format
      properties:
        timestamp:
          $ref: '#/components/schemas/DateTime'
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: HTTP status reason phrase
        message:
          type: string
          description: Human-readable error description
        path:
          type: string
          description: The relative URI of the request that caused the error.
          example: "/api/v1/resource-path"
        error_code:
          type: string
          description: Application-specific error code
          enum:
            - RESOURCE_NOT_FOUND
            - ILLEGAL_STATE
            - SECURITY_EXCEPTION
            - PAYLOAD_NOT_AS_EXPECTED
            - EMAIL_HEALTH_CHECK_FAILED
            - MISSING_REQUIRED_PART
            - DUPLICATE_EMAIL
            - DUPLICATE_SUPABASE_ID
            - DIRECTORY_UPLOAD_FAILED
            - AUTHORIZATION_DENIED
            - ENTITY_NOT_FOUND
            - MISSING_REQUEST_HEADER
            - MISSING_AUTH_CREDS
            - BAD_FORM
            - MISSING_AUTHORIZATION
            - INTERNAL_SERVER_ERROR
            - MISSING_REQUIRED_PARAMETER
            - INVALID_PARAMETER
            - MALFORMED_JSON
            - METHOD_NOT_ALLOWED
        details:
          type: object
          description: Additional error context
  responses:
    InternalServer:
      description: Internal Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request parameters or payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions for requested operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Resource conflict or duplicate entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'